Add Shard Shrikes
Batty monsters that spit 3d30 Throw Icicle, resist cold, and flock
in small bands. Intended to add a little variety and danger to
Cocytus, where they live.

Credit to Lasty for the name and, in a sense, the inspiration.
Credit to Hangedman for spell choice & habitat; thanks to others
as appropriate.o