Make butchering/bottling single-turn.
The only objections raised were that it might be too strong versus
corpse-raising enemies, but the consensus was that a turn spent, and
the necessity to be on the square, were enough of a cost. The other
was that it might make it too easy to switch blood level very quickly
as a vampire, but carrying lots of potions usually does the same thing.